refactor: DSV4 family owns perf data (AIC-1095)
Phase 3.5 — moves the 4 DeepSeek-V4 op classes
(DeepSeekV4MHCModule, _BaseDeepSeekV4AttentionModule,
ContextDeepSeekV4AttentionModule, GenerationDeepSeekV4AttentionModule)
out of ``operations/_legacy.py`` into a new ``operations/dsv4.py``.

Data ownership follows the Lazy-load Pattern A established by Phase 2:
each op class owns its ``_data_cache`` keyed by the canonical 5-tuple
``(systems_root, system, backend, version, enable_shared_layer)``;
``load_data(database)`` is idempotent and binds the legacy instance
attributes for downstream consumers. The context class owns three
caches — the merged ctx table, a deepcopy used for topk piecewise
lookup, and the sparse-kernel sidecar dict (paged_mqa_logits +
hca_attn).

Helpers moved alongside the ops:
- ``_dsv4_robust_3d_lookup`` (exact → cubic → sampled-batch fallback)
- ``_deep_merge_dsv4_dicts`` (csa+hca split-file merge)
- ``_deepseek_v4_attention_sol`` (shared SOL formula)
- ``_lookup_sparse_kernel`` (kernel-Δ for prefix context)
- ``_load_dsv4_split`` (init-time merger)

The two module-level helpers (``_dsv4_robust_3d_lookup`` and
``_deep_merge_dsv4_dicts``) are re-exported lazily from
``perf_database`` via ``__getattr__`` so tests importing them from
that path keep working without triggering ``operations/__init__.py``
at perf_database import time (avoids a circular import via
``_legacy.py``).

``PerfDatabase.query_mhc_module`` /
``query_context_deepseek_v4_attention_module`` /
``query_generation_deepseek_v4_attention_module`` reduce to one-line
delegations to the new classmethods.
``PerfDatabase._lookup_dsv4_sparse_kernel`` is kept as a thin
forwarding wrapper because tests invoke it via the unbound form
``PerfDatabase._lookup_dsv4_sparse_kernel(db, ...)``; ISSUE-16 cleanup
retires this wrapper once those tests migrate.

CSV loaders (``load_context_dsv4_kind_module_data``,
``load_generation_dsv4_kind_module_data``,
``load_dsv4_sparse_kernel_data``, ``load_mhc_module_data``) stay in
``perf_database.py`` so existing test imports work unchanged — same
pattern as the GEMM / Attention migrations.

Eager ``OpClass.load_data(self)`` calls added in
``PerfDatabase.__init__`` alongside the existing GEMM /
ContextAttention / GenerationAttention transition shims.
ISSUE-16 cleanup retires all of them.

Verification:
- 670 unit tests pass + 1 skipped (same as baseline)
- 38 regression snapshots bit-identical + 6 skipped (same as baseline)
- ruff check + format clean
