+++ src/aiconfigurator/sdk/operations/__init__.py
@@ -25,16 +25,12 @@ from __future__ import annotations
 # logger lives in ``_legacy.py``; per-family migrations don't redirect their
 # log calls until their owning issue lands.
 from aiconfigurator.sdk.operations._legacy import (
-    ContextDeepSeekV4AttentionModule,
-    DeepSeekV4MHCModule,
     FallbackOp,
-    GenerationDeepSeekV4AttentionModule,
     MoE,
     MoEDispatch,
     OverlapOp,
     TrtLLMWideEPMoE,
     TrtLLMWideEPMoEDispatch,
-    _BaseDeepSeekV4AttentionModule,
     logger,  # noqa: F401
 )
 
@@ -46,6 +42,12 @@ from aiconfigurator.sdk.operations.attention import ContextAttention, Generation
 from aiconfigurator.sdk.operations.base import Operation, clear_all_op_caches
 from aiconfigurator.sdk.operations.communication import NCCL, P2P, CustomAllReduce
 from aiconfigurator.sdk.operations.dsa import ContextDSAModule, GenerationDSAModule
+from aiconfigurator.sdk.operations.dsv4 import (
+    ContextDeepSeekV4AttentionModule,
+    DeepSeekV4MHCModule,
+    GenerationDeepSeekV4AttentionModule,
+    _BaseDeepSeekV4AttentionModule,
+)
 from aiconfigurator.sdk.operations.elementwise import ElementWise
 from aiconfigurator.sdk.operations.embedding import Embedding
 from aiconfigurator.sdk.operations.gemm import GEMM
